dashboard/requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.11.0
pandas==2.2.0
plotly==5.20.0
